fix wrong field name when parent is null
