+++ imports/api/query-builder/index.js
@@ -1,5 +1,4 @@
 export {default as queryBuilder, makeExpression, validateConditions} from './polish-notation';
 export {default as aggsBuilder} from './aggregations';
-export {default as builder} from './query-builder';
 export {default as Query} from './query';
-export {default as Condition} from './condition';
+export {default as Condition} from './single-query';
